close #9: RAM-friendly data representation
Before this commit, charming was backed by data.json.lzo.64, a single
1114112-array of objects, each of which contains the information that
we need to know when the user chooses the corresponding code point. A
couple of examples (for U+231A and U+9D97) are shown below:

	{
		"age": "Unicode 1.1",
		"gc": "Other Symbol (So)",
		"name": "WATCH",
		"block": "Miscellaneous Technical",
		"emoji": 3
	}

	{
		"age": "Unicode 1.1",
		"han": true,
		"name": "a kind of hawk; cuckoo",
		"block": "CJK Unified Ideographs",
		"mpy": "t\u00ed"
	}

According to Chrome/64 and Firefox/57, the 25995487 bytes of JSON were
being expanded to at least four times that size when parsed. No wonder
we were reliably crashing on the iPhone SE (iOS/11.2.6).

The new representation breaks data.json.lzo.64 into seven files that
exploit three observations: strings can be used as immutable arrays of
unsigned 16-bit integers [ECMA-262 #3 §4.3.16] that should only waste
O(k) memory for any k strings of any length, the details we currently
associate with each code point are strings and Boolean values, and
there are fewer than 65535 unique strings (49872 to be exact).

One copy of each string is written to the “string” file, the Boolean
values are packed into a 557056-string and written to the “bits” file,
and each of the remaining files represents a 1114112-string of indices
into the string pool (where 0xFFFF means that none are applicable).

*.json.lzo.64/*.json was: 1749168/25995487
*.json.lzo.64/*.json now: 1592188/38374860
*.json.lzo.64/*.json: −8.9%/+47.7%

Chrome/64 now: 107 MB
Chrome/64 now: 21 MB
Chrome/64: −80.3%

Firefox/57 was: 156 MB
Firefox/57 now: 20 MB
Firefox/57: −87.1%