fourth biennial data optimisation
We first optimised the system in b42725d2c7abe by compiling the UCD
down to a JSON file with only the details we need ahead of time.

Then in d9ccb87e22561 we started compressing that file with LZO1X and
decompressing it in the client script. HTTP compression would probably
have been better for the client’s CPU time, but this worked regardless
of web server configuration.

Then in 2c2e92f10e7cd we switched from AoS to SoA, introduced a string
pool with one packed string of indices for each field, and moved the
remaining flags to an additional packed string.

By that point, our data format was very complicated. We were packing
integers into JavaScript strings, which in turn we delivered in JSON,
which in turn we compressed with LZO1X, which in turn we btoa-encoded
to avoid binary cleanliness problems.

This commit replaces each of those with a binary file, compressed with
Content-Encoding: br, delivered by the Fetch API in an ArrayBuffer, an
approach which consumes the least network activity and client CPU time
ever, but should keep client RAM usage at current levels.