upgrade babel-loader, css-loader, file-loader, html-loader, sass-loader, style-loader and worker-loader to latest
due to their use in old.html, webpack now checks that the files
jquery-1.8.3.min.js and old.css exist - and they don't, at least not in
src/. this previously worked because both files lived in dist/ and thus
would be in the right place after a build.

we move both to src/, which necessitates configuring prettier to ignore
them too.

additionally, file-loader does not work for custom fonts in webpack ^5,
leading to ots parsing errors (fonts are now text files, webpack goes
sideways).

we switch just the fonts (for now) to the asset/resource modules.