upgrade @types/react, @types/react-dom, @types/react-window, clipboard-polyfill, grapheme-iterator, react, react-dom, react-use, react-virtualized-auto-sizer, react-window and regenerator-runtime; remove @fortawesome/fontawesome-free
**--legacy-peer-deps and NODE_OPTIONS=--openssl-legacy-provider are now
no longer required**

we remove fontawesome since we use exactly one icon (fa-brands
fa-github) which isn't very compatible with the font subsetting which
was previously performed. to include the styles for .fa-github, we must
import ~@fortawesome/fontawesome-free/scss/brands in new.scss. the
brands file is, annoyingly, quite different from
~@fortawesome/fontawesome-free/scss/icons (which was already imported,
and functioned correctly after the upgrade).

icons simply iterates through the list of icons and generates styles for each:
    @each $name, $icon in $fa-icons {
      .#{$fa-css-prefix}-#{$name}::before { content: unquote("\"#{ $icon }\""); }
    }

whereas brands also tries to pull in the brands font itself.
    @font-face {
      font-family: 'Font Awesome 6 Brands';
      font-style: normal;
      font-weight: 400;
      font-display: $fa-font-display;
      src: url('#{$fa-font-path}/fa-brands-400.woff2') format('woff2'),
        url('#{$fa-font-path}/fa-brands-400.ttf') format('truetype');
    }
    ...
    @each $name, $icon in $fa-brand-icons {
      .#{$fa-css-prefix}-#{$name}:before { content: unquote("\"#{ $icon }\""); }
    }

this don't work so well for us.

we replace the github icon with the svg of the fontawesome github icon.