upgrade dependencies (#50)
* update package-lock.json with 'npm install --legacy-peer-deps' on node v20.11.0, npm 10.2.4

running 'npm install' from a fresh state /works/, but running it again
gives an error about conflicting peer dependencies.

    npm ERR! code ERESOLVE
    npm ERR! ERESOLVE could not resolve
    npm ERR!
    npm ERR! While resolving: react-use@9.4.1
    npm ERR! Found: react@18.0.0
    ...
    npm ERR!
    npm ERR! Could not resolve dependency:
    npm ERR! peer react@"^16.8.0" from react-use@9.4.1
    ...
    npm ERR!
    npm ERR! Conflicting peer dependency: react@16.14.0
    npm ERR! node_modules/react
    ...
    npm ERR!
    npm ERR! Fix the upstream dependency conflict, or retry
    npm ERR! this command with --force or --legacy-peer-deps
    npm ERR! to accept an incorrect (and potentially broken) dependency resolution.

the '--legacy-peer-deps' flag brings back behaviour from older versions
of npm (https://stackoverflow.com/a/66620869) which makes for our
upgrade use case. with the flag, both the initial install and a second
unnecessary install succeed.

* upgrade webpack, webpack-cli, webpack-dev-server, compression-webpack-plugin and html-webpack-plugin to latest

adding the <head> tag to new.html, although regrettable, is required to
avoid webpack inserting a head element before the actual content, and
importantly, the doctype declaration.

dist/new.html without <head>:
    <head><script defer src="120a0a86d9a09259557c/new.js"></script></head><!DOCTYPE HTML>
    <meta charset="UTF-8">
    <title>charming</title>
    <meta name="viewport" content="width=device-width, user-scalable=no">
    <meta name="google" value="notranslate"><!-- inhibit chrome automatic translation -->
    <main><div class="Charming"><div class="Detail"></div><div class="Map measurer"></div></div>

dist/new.html with <head>:
    <!DOCTYPE HTML>
    <head>
    <meta charset="UTF-8">
    <title>charming</title>
    <meta name="viewport" content="width=device-width, user-scalable=no">
    <meta name="google" value="notranslate"><!-- inhibit chrome automatic translation -->
    <script defer src="2532c964e0b46578dcb6/new.js"></script></head>
    <main><div class="Charming"><div class="Detail"></div><div class="Map measurer"></div></div>

* upgrade babel-loader, css-loader, file-loader, html-loader, sass-loader, style-loader and worker-loader to latest

due to their use in old.html, webpack now checks that the files
jquery-1.8.3.min.js and old.css exist - and they don't, at least not in
src/. this previously worked because both files lived in dist/ and thus
would be in the right place after a build.

we move both to src/, which necessitates configuring prettier to ignore
them too.

additionally, file-loader does not work for custom fonts in webpack ^5,
leading to ots parsing errors (fonts are now text files, webpack goes
sideways).

we switch just the fonts (for now) to the asset/resource modules.

* upgrade prettier to latest

* upgrade @babel/core, @babel/preset-env, @babel/preset-react, @babel/preset-typescript, @types/jest, babel-jest, jest, sass and typescript to latest

sass has deprecated using `/` for division
(https://sass-lang.com/d/slash-div) which is easy to fix in new.sass.
however, until the frontend dependencies are updated, sass complains
about some vendor code.

in the webpack config, we disable warnings for anything in node_modules
temporarily.

* remove html-loader

html-loader is only used when a script imports a html file. we don't do
this (as all the html entrypoints are managed with html-webpack-plugin
providing HtmlPlugin), so we don't need it.

* move .bin files to asset modules

* upgrade @types/react, @types/react-dom, @types/react-window, clipboard-polyfill, grapheme-iterator, react, react-dom, react-use, react-virtualized-auto-sizer, react-window and regenerator-runtime; remove @fortawesome/fontawesome-free

**--legacy-peer-deps and NODE_OPTIONS=--openssl-legacy-provider are now
no longer required**

we remove fontawesome since we use exactly one icon (fa-brands
fa-github) which isn't very compatible with the font subsetting which
was previously performed. to include the styles for .fa-github, we must
import ~@fortawesome/fontawesome-free/scss/brands in new.scss. the
brands file is, annoyingly, quite different from
~@fortawesome/fontawesome-free/scss/icons (which was already imported,
and functioned correctly after the upgrade).

icons simply iterates through the list of icons and generates styles for each:
    @each $name, $icon in $fa-icons {
      .#{$fa-css-prefix}-#{$name}::before { content: unquote("\"#{ $icon }\""); }
    }

whereas brands also tries to pull in the brands font itself.
    @font-face {
      font-family: 'Font Awesome 6 Brands';
      font-style: normal;
      font-weight: 400;
      font-display: $fa-font-display;
      src: url('#{$fa-font-path}/fa-brands-400.woff2') format('woff2'),
        url('#{$fa-font-path}/fa-brands-400.ttf') format('truetype');
    }
    ...
    @each $name, $icon in $fa-brand-icons {
      .#{$fa-css-prefix}-#{$name}:before { content: unquote("\"#{ $icon }\""); }
    }

this don't work so well for us.

we replace the github icon with the svg of the fontawesome github icon.

* upgrade corejs to latest

not sure how i forgot this one

* fix TS2315 in @types/ws by overriding @types/node

* fix breakage in state.test.ts since Jest 27 and 28

* move back to caret deps; not sure exact deps are actually better

---------

Co-authored-by: Delan Azabani <delan@azabani.com>